Add setting dialog, menu items and manage large files
settings_dlg gathers paths for rs274, tooltable and temporary gcode file.
If any of these are not set, pops up at launch and sets them
in variables within the class.

File Open and File Save As added

Large file management by setting limits to initial amount of a file read into
QGCodeEditor widget, with further blocks read in if cursor position nears
the end of the current amount in the editor.

This reduces the time to load a 10,000 line file, from 30 seconds just for the editor
to +/- 5secs for both editor and GLViewer combined

This is an interim measure.  Full solution will be via changes to editor itself
to do this irrespective of the input source.

Signed-off-by: Mick <arceye@mgware.co.uk>